[Bronze II] Title: 숫자, Time: 24 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -3,15 +3,15 @@
 using namespace std;
 
 int main() {
-	int A, B;
-	int bigger, smaller;
+	long long A, B;
+	long long bigger, smaller;
 
 	cin >> A >> B;
 	if (A > B) {
 		bigger = A;
 		smaller = B;
 		cout << A - B - 1 << "\n";
-		for (int i = B+1; i < A; i++) {
+		for (long long i = B+1; i < A; i++) {
 			cout << i << " ";
 		}
 	}
@@ -19,7 +19,7 @@ int main() {
 		bigger = B;
 		smaller = A;
 		cout << B - A - 1 << "\n";
-		for (int i = A+1; i < B; i++) {
+		for (long long i = A+1; i < B; i++) {
 			cout << i << " ";
 		}
 	}
